requirements.txt
groq
python-dotenv
streamlit
scikit-learn
numpy
pandas
openai
python-dotenv
